v4.5.3 update fixes play count zeroing bug
git-svn-id: https://plugins.svn.wordpress.org/video-embed-thumbnail-generator/tags/4.5.3@1197154 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ video-embed-thumbnail-generator.php
@@ -3,10 +3,11 @@
 Plugin Name: Video Embed & Thumbnail Generator
 Plugin URI: http://www.kylegilman.net/2011/01/18/video-embed-thumbnail-generator-wordpress-plugin/
 Description: Generates thumbnails, HTML5-compliant videos, and embed codes for locally hosted videos. Requires FFMPEG or LIBAV for encoding.
-Version: 4.5
+Version: 4.5.3
 Author: Kyle Gilman
 Author URI: http://www.kylegilman.net/
 Text Domain: video-embed-thumbnail-generator
+Domain Path: /languages
 
 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -58,7 +59,7 @@ function kgvid_default_options_fn() {
 	$edit_others_capable = kgvid_check_if_capable('edit_others_posts');
 
 	$options = array(
-		"version" => 4.5,
+		"version" => 4.501,
 		"embed_method" => "Video.js",
 		"jw_player_id" => "",
 		"template" => false,
@@ -268,7 +269,12 @@ function kgvid_get_attachment_meta($post_id) {
 				delete_post_meta($post_id, "_kgflashmediaplayer-".$key);
 			}
 
-			array_filter($kgvid_postmeta, 'strlen'); //remove empty elements
+			foreach ( $kgvid_postmeta as $key => $value ) {
+				if ( $value === null ) {
+					unset( $kgvid_postmeta[ $meta ] ); //remove empty elements
+				}
+			}
+
 			kgvid_save_attachment_meta($post_id, $kgvid_postmeta);
 
 		}
@@ -283,9 +289,11 @@ function kgvid_get_attachment_meta($post_id) {
 
 function kgvid_save_attachment_meta($post_id, $kgvid_postmeta) {
 
-	$options = kgvid_get_options();
-
 	if ( is_array($kgvid_postmeta) ) {
+
+		$options = kgvid_get_options();
+		$kgvid_old_postmeta = kgvid_get_attachment_meta($post_id);
+		$kgvid_postmeta = array_merge($kgvid_old_postmeta, $kgvid_postmeta); //make sure all keys are saved
 
 		foreach ( $kgvid_postmeta as $key => $meta ) { //don't save if it's the same as the default values or empty
 
@@ -1979,8 +1987,16 @@ function KGVID_shortcode($atts, $content = ''){
 
 				if ( !empty($id) ) { //if the video is an attachment in the WordPress db
 
-					$content = wp_get_attachment_url($id);
-					if ( $content == false ) { echo "Invalid video ID"; continue; }
+					$attachment_url = wp_get_attachment_url($id);
+					if ( $attachment_url == false ) { echo "Invalid video ID"; continue; }
+					$exempt_cdns = array('amazonaws.com', 'rackspace.com', 'netdna-cdn.com', 'nexcess-cdn.net', 'limelight.com'); //don't replace URLs that point to CDNs
+					$exempt_url = false;
+					foreach ( $exempt_cdns as $exempt_cdn ) {
+						if ( strpos($content, $exempt_cdn) !== false ) {
+							$exempt_url = true;
+						}
+					}
+					if ( !$exempt_url ) { $content = $attachment_url; }
 
 					$encodevideo_info = kgvid_encodevideo_info($content, $id);
 					$attachment_info = get_post( $id );
@@ -1997,7 +2013,7 @@ function KGVID_shortcode($atts, $content = ''){
 					if ( !empty($poster_id) ) {
 						$poster_image_src = wp_get_attachment_image_src($poster_id, 'full');
 						$query_atts['poster'] = $poster_image_src[0];
-						if ( intval($query_atts['width']) <= get_option('medium_size_h') ) {
+						if ( strpos($query_atts['width'], '%') === false && intval($query_atts['width']) <= get_option('medium_size_h') ) {
 							$query_atts['poster'] = kgvid_get_attachment_medium_url($poster_id);
 						}
 					}
@@ -2106,8 +2122,7 @@ function KGVID_shortcode($atts, $content = ''){
 
 				} //if Strobe Media
 
-				//wp_localize_script( 'kgvid_video_embed', 'kgvid_video_vars_'.$div_suffix, $video_variables ); //add video variables in footer
-
+				$video_variables = str_replace('&nbsp;', ' ', $video_variables);
 				$code .= '<div id="kgvid_'.$div_suffix.'_wrapper" class="kgvid_wrapper'.$aligncode.'">'."\n\t\t\t";
 				$code .= '<div id="video_'.$div_suffix.'_div" class="fitvidsignore kgvid_videodiv" data-id="'.$div_suffix.'" data-kgvid_video_vars="'.esc_attr(json_encode($video_variables)).'" itemprop="video" itemscope itemtype="https://schema.org/VideoObject">';
 				if ( $query_atts["poster"] != '' ) { $code .= '<meta itemprop="thumbnailUrl" content="'.esc_attr($query_atts["poster"]).'" />'; }
@@ -2138,7 +2153,7 @@ function KGVID_shortcode($atts, $content = ''){
 							if ( empty($value) ) { $track_attribute[$attribute] = $query_atts['track_'.$attribute]; }
 						}
 						if ( $options['embed_method'] == "WordPress Default" && $track_attribute['kind'] == 'captions' ) { $track_attribute['kind'] = 'subtitles'; }
-						$track_code .= "\t\t\t\t\t<track kind='".esc_attr($track_attribute['kind'])."' src='".esc_attr($track_attribute['src'])."' srclang='".esc_attr($track_attribute['srclang'])."' label='".esc_attr($track_attribute['label'])."' ".esc_attr($track_attribute['default'])." />\n";
+						$track_code .= "\t\t\t\t\t<track id='".$div_suffix."_text_".$track."' kind='".esc_attr($track_attribute['kind'])."' src='".esc_attr($track_attribute['src'])."' srclang='".esc_attr($track_attribute['srclang'])."' label='".esc_attr($track_attribute['label'])."' ".esc_attr($track_attribute['default'])." />\n";
 					}
 				}
 
@@ -4870,7 +4885,7 @@ function kgvid_image_attachment_fields_to_edit($form_fields, $post) {
 				URL: <input name="attachments['. $post->ID .'][kgflashmediaplayer-track]['.$track.'][src]" id="attachments-'. $post->ID .'-kgflashmediaplayer-track_'.$track.'_src" type="text" value="'.$kgvid_postmeta['track'][$track]['src'].'" class="text" style="width:180px;"><br />
 				'._x('Language code:', 'two-letter code indicating track\'s language', 'video-embed-thumbnail-generator').' <input name="attachments['. $post->ID .'][kgflashmediaplayer-track]['.$track.'][srclang]" id="attachments-'. $post->ID .'-kgflashmediaplayer-track_'.$track.'_srclang" type="text" value="'.$kgvid_postmeta['track'][$track]['srclang'].'" maxlength="2" style="width:40px;"><br />
 				'.__('Label:', 'video-embed-thumbnail-generator').' <input name="attachments['. $post->ID .'][kgflashmediaplayer-track]['.$track.'][label]" id="attachments-'. $post->ID .'-kgflashmediaplayer-track_'.$track.'_label" type="text" value="'.$kgvid_postmeta['track'][$track]['label'].'" class="text" style="width:172px;"><br />
-				'.__('Default:', 'video-embed-thumbnail-generator').'<input '.checked(array_key_exists('default', $kgvid_postmeta['track'][$track]), true, false).' name="attachments['. $post->ID .'][kgflashmediaplayer-track]['.$track.'][default]" id="attachments-'. $post->ID .'-kgflashmediaplayer-track_'.$track.'_default" type="checkbox" value="default"></div>';
+				'.__('Default:', 'video-embed-thumbnail-generator').'<input '.checked($kgvid_postmeta['track'][$track]['default'], 'default', false).' name="attachments['. $post->ID .'][kgflashmediaplayer-track]['.$track.'][default]" id="attachments-'. $post->ID .'-kgflashmediaplayer-track_'.$track.'_default" type="checkbox" value="default"></div>';
 			}
 		}
 
@@ -5259,11 +5274,26 @@ function kgvid_video_attachment_fields_to_save($post, $attachment) {
 					update_post_meta($post['ID'], '_kgflashmediaplayer-poster-id', $thumb_id);
 				}
 			}
-			else {
+
+			if ( empty($thumb_url) ) {
 				delete_post_meta($post['ID'], '_kgflashmediaplayer-poster');
 				delete_post_meta($post['ID'], '_kgflashmediaplayer-poster-id');
+				delete_post_thumbnail($post['ID']);
 			}
-			update_post_meta($post['ID'], '_kgflashmediaplayer-poster', $thumb_url);
+			else {
+
+				update_post_meta($post['ID'], '_kgflashmediaplayer-poster', $thumb_url);
+
+				if ( empty($thumb_id) ) { //we're not saving a new thumbnail
+					$poster_id = get_post_meta($post['ID'], '_kgflashmediaplayer-poster-id', true);
+					if ( empty($poster_id) ) { //the poster_id meta was accidentally deleted
+						$thumb_url_id = kgvid_url_to_id($thumb_url);
+						if ( $thumb_url_id ) { update_post_meta($post['ID'], '_kgflashmediaplayer-poster-id', $thumb_url_id); }
+					}
+				}
+
+			}
+
 		}
 
 		if( isset($attachment['kgflashmediaplayer-featured']) ) {
@@ -5299,6 +5329,12 @@ function kgvid_video_attachment_fields_to_save($post, $attachment) {
 		$checkboxes = array( 'lockaspect', 'featured', 'showtitle', 'downloadlink' ); //make sure unchecked checkbox values are saved
 		foreach ( $checkboxes as $checkbox ) {
 			if( !isset($attachment['kgflashmediaplayer-'.$checkbox]) ) { $attachment['kgflashmediaplayer-'.$checkbox] = "false"; }
+		}
+
+		if ( isset( $attachment['kgflashmediaplayer-track'] ) && is_array( $attachment['kgflashmediaplayer-track'] ) ) {
+			foreach ( $attachment['kgflashmediaplayer-track'] as $track => $track_attribute ) {
+				if ( !array_key_exists('default', $track_attribute) ) { $attachment['kgflashmediaplayer-track'][ $track ]['default'] = ''; }
+			}
 		}
 
 		$kgvid_postmeta = array();
